requirements.txt
wget
nest-asyncio
llama-index
llama-index-embeddings-fastembed
llama-index-vector-stores-qdrant
llama-index-llms-groq
streamlit
python-dotenv
fastembed
lxml
qdrant_client
unstructured 
streamlit_drawable_canvas